Setup a stripe subscription payment provider
